Resolving int too big for byte conversion
For the current version of the code itself, if you run it, you'll recieve an error that the int value is too big to be converted to a byte. I resolved the issue to an extent by limiting the range of the int value being converted (int_data = int_data % 65536 after line 80), but it still results in a jpg that is still identifiable from the original.